+++ bin/nanocld/src/utils/stream.rs
@@ -22,6 +22,7 @@ where
       status: StatusCode::INTERNAL_SERVER_ERROR,
       msg: format!("Failed to serialize stream item: {err}"),
     })?;
+    println!("item: {}", item);
     Ok(Bytes::from(item + "\r\n"))
   })
 }