+++ crates/nanocl_stubs/src/metric.rs
@@ -1,22 +1,25 @@
 #[cfg(feature = "serde")]
 use serde::{Serialize, Deserialize};
 
-/// Generic namespace query filter
-#[derive(Debug, Clone)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "serde", serde(rename_all = "PascalCase"))]
-pub struct GenericNspQuery {
-  /// Name of the namespace
-  pub namespace: Option<String>,
-}
-
-/// Generic delete response
+/// Metric entry
 #[derive(Debug)]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[cfg_attr(feature = "serde", serde(rename_all = "PascalCase"))]
-#[cfg_attr(feature = "dev", derive(ToSchema))]
-pub struct GenericDelete {
-  /// Number of deleted items
-  pub count: usize,
+pub struct Metric {
+  pub key: uuid::Uuid,
+  pub created_at: chrono::NaiveDateTime,
+  pub expire_at: chrono::NaiveDateTime,
+  pub node_name: String,
+  pub kind: String,
+  pub data: serde_json::Value,
+}
+
+/// Filter metrics query
+#[derive(Debug)]
+#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", serde(rename_all = "PascalCase"))]
+pub struct MetricFilterQuery {
+  pub kind: String,
 }