+++ crates/nanocld_client/src/cargo.rs
@@ -462,6 +462,8 @@ impl NanocldClient {
       )
       .await?;
 
+    println!("res: {:?}", res);
+
     Ok(Self::res_stream(res).await)
   }