+++ scripts/install_dev_image.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 ## name: build_dev_image.sh
 
-docker pull cockroachdb/cockroach:v22.2.6
+docker pull cockroachdb/cockroach:v22.2.7
 docker pull ghcr.io/nxthat/metrsd:0.2.0
 docker pull nexthat/nanocl-get-started:latest
 docker pull ghcr.io/nxthat/nanocl-dev:dev